django_task_tracker-master/requirements.txt
asgiref==3.3.1
automium==0.2.6
automium-web==0.1.1
beautifulsoup4==4.10.0
certifi==2021.10.8
charset-normalizer==2.0.7
click==8.0.3
colorama==0.4.4
cycler==0.10.0
Django==2.2
django-crispy-forms==1.11.2
django-filter==21.1
djangorestframework==3.13.1
Flask==2.0.2
idna==3.3
itsdangerous==2.0.1
Jinja2==3.0.3
joblib==1.1.0
kiwisolver==1.3.2
legacy==0.1.6
MarkupSafe==2.0.1
matplotlib==3.4.3
news==1.0
numpy==1.21.2
pandas==1.3.4
Pillow==8.1.0
protobuf==3.7.1
pycodestyle==2.5.0
pyparsing==3.0.1
python-dateutil==2.8.2
pytz==2019.1
requests==2.26.0
scikit-learn==1.0.1
scipy==1.7.1
seaborn==0.11.2
six==1.12.0
soupsieve==2.3.1
sqlparse==0.3.0
threadpoolctl==3.0.0
urllib3==1.26.7
Werkzeug==2.0.2
